moved more logic back to directus sdk, deprecated all custom modules, moved more adapters and logic to backups directory

+++ layers/social/app/composables/content/useDirectusRequest.ts
@@ -1,7 +1,8 @@
 declare function useNuxtApp(): any
 
+import { getAssetURL } from '#shared/app/utils/get-asset-url'
+
 export default function useDirectusRequest() {
-  const nuxt = typeof useNuxtApp === 'function' ? useNuxtApp() as any : (globalThis as any).__nuxtApp || {}
 
   function _toast(message: string) {
     try {
@@ -10,15 +11,20 @@ export default function useDirectusRequest() {
     } catch (_) {}
   }
 
-  function resolveSdkContent() {
-    return nuxt?.$sdk?.content || null
+  function resolveDirectusClient() {
+    return nuxt?.$directus || (globalThis as any).__directus || null
+  }
+
+  function resolveBuilder(name: string) {
+    return nuxt?.[`$${name}`] || (globalThis as any)[`__${name}`] || null
   }
 
   async function request(config: any) {
     try {
-      const sdkContent = resolveSdkContent()
-      if (sdkContent && typeof sdkContent.request === 'function') return await sdkContent.request(config)
-
+      const client = resolveDirectusClient()
+      if (client && typeof client.request === 'function') {
+        return await client.request(config)
+      }
       _toast('Data client unavailable')
       throw new Error('Data client.request is not available')
     } catch (e) {
@@ -28,65 +34,78 @@ export default function useDirectusRequest() {
   }
 
   async function readItems(collection: string, opts?: any) {
-    const sdkContent = resolveSdkContent()
-    if (sdkContent && typeof sdkContent.readItems === 'function') return await sdkContent.readItems(collection, opts)
-
+    const client = resolveDirectusClient()
+    const builder = resolveBuilder('readItems')
+    if (client && typeof client.request === 'function' && typeof builder === 'function') {
+      return await client.request(builder(collection, opts))
+    }
     _toast('readItems not available')
     throw new Error('readItems not available')
   }
 
   async function readItem(collection: string, id: string, opts?: any) {
-    const sdkContent = resolveSdkContent()
-    if (sdkContent && typeof sdkContent.readItem === 'function') return await sdkContent.readItem(collection, id, opts)
-
+    const client = resolveDirectusClient()
+    const builder = resolveBuilder('readItem')
+    if (client && typeof client.request === 'function' && typeof builder === 'function') {
+      return await client.request(builder(collection, id, opts))
+    }
     _toast('readItem not available')
     throw new Error('readItem not available')
   }
 
   async function readFieldsByCollection(collection: string, opts?: any) {
-    const sdkContent = resolveSdkContent()
-    if (sdkContent && typeof sdkContent.readFieldsByCollection === 'function') return await sdkContent.readFieldsByCollection(collection, opts)
-
+    const client = resolveDirectusClient()
+    const builder = resolveBuilder('readFieldsByCollection')
+    if (client && typeof client.request === 'function' && typeof builder === 'function') {
+      return await client.request(builder(collection, opts))
+    }
     _toast('readFieldsByCollection not available')
     throw new Error('readFieldsByCollection not available')
   }
 
   async function deleteItem(collection: string, id: string) {
-    const sdkContent = resolveSdkContent()
-    if (sdkContent && typeof sdkContent.deleteItem === 'function') return await sdkContent.deleteItem(collection, id)
-
+    const client = resolveDirectusClient()
+    const builder = resolveBuilder('deleteItem')
+    if (client && typeof client.request === 'function' && typeof builder === 'function') {
+      return await client.request(builder(collection, id))
+    }
     _toast('deleteItem not available')
     throw new Error('deleteItem not available')
   }
 
   async function createItem(collection: string, data: any) {
-    const sdkContent = resolveSdkContent()
-    if (sdkContent && typeof sdkContent.createItem === 'function') return await sdkContent.createItem(collection, data)
-
+    const client = resolveDirectusClient()
+    const builder = resolveBuilder('createItem')
+    if (client && typeof client.request === 'function' && typeof builder === 'function') {
+      return await client.request(builder(collection, data))
+    }
     _toast('createItem not available')
     throw new Error('createItem not available')
   }
 
   async function updateItem(collection: string, idOrFilter: any, data: any) {
-    const sdkContent = resolveSdkContent()
-    if (sdkContent && typeof sdkContent.updateItem === 'function') return await sdkContent.updateItem(collection, idOrFilter, data)
-
+    const client = resolveDirectusClient()
+    const builder = resolveBuilder('updateItem')
+    if (client && typeof client.request === 'function' && typeof builder === 'function') {
+      return await client.request(builder(collection, idOrFilter, data))
+    }
     _toast('updateItem not available')
     throw new Error('updateItem not available')
   }
 
   async function uploadFiles(formData: FormData) {
-    const sdkContent = resolveSdkContent()
-    if (sdkContent && typeof sdkContent.uploadFiles === 'function') return await sdkContent.uploadFiles(formData)
-
+    const client = resolveDirectusClient()
+    const builder = resolveBuilder('uploadFiles')
+    if (client && typeof client.request === 'function' && typeof builder === 'function') {
+      return await client.request(builder(formData))
+    }
     _toast('uploadFiles not available')
     throw new Error('uploadFiles not available')
   }
 
   async function getAssetUrl(file: any) {
-    const sdkMedia = nuxt?.$sdk?.media
-    if (sdkMedia && typeof sdkMedia.getAssetUrl === 'function') return sdkMedia.getAssetUrl(file)
-    return ''
+    const url = getAssetURL(file)
+    return typeof url === 'string' ? url : ''
   }
 
   return { request, readItems, readItem, readFieldsByCollection, createItem, updateItem, deleteItem, uploadFiles, getAssetUrl }
